requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-session==0.5.0
python-dotenv==1.0.1
pdfkit==1.0.0
fpdf2==2.8.2

langchain==0.1.17
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-community>=0.0.36,<0.1
langchain-text-splitters==0.0.2
langgraph==0.0.26

openai==1.69.0
